server: get valid DRM formats at startup
Not sure of the performance implications of calling
eglQueryDmaBufFormatsEXT and eglQueryDmaBufModifiersEXT (I was too lazy
to benchmark either) but it's nice to have them queried ahead of time to
be able to print them out for debugging purposes regardless.

At some point I should probably just cave and link libdrm. It will
probably finally need to happen whenever I get around to supporting
explicit sync in the DMABUF importing code.